saving stuff - forced computer update
