+++ pennylane/ftqc/ftqc_device.py
@@ -15,12 +15,8 @@
 Contains an implementation of a PennyLane frontend (Device) for an FTQC/MBQC based
 hardware device (or emulator)
 """
-import copy
 from dataclasses import replace
-from functools import cached_property
 from pathlib import Path
-
-import numpy as np
 
 import pennylane as qml
 from pennylane.devices.capabilities import DeviceCapabilities, validate_mcm_method
@@ -38,11 +34,12 @@ from pennylane.ftqc import (
     convert_to_mbqc_gateset,
     diagonalize_mcms,
 )
-from pennylane.measurements import Shots
 from pennylane.ops import RZ
 from pennylane.tape.qscript import QuantumScript
 from pennylane.transforms import combine_global_phases, split_non_commuting
 from pennylane.typing import Result
+
+from .quantum_script_sequence import QuantumScriptSequence
 
 
 class FTQCQubit(Device):
@@ -89,7 +86,6 @@ class FTQCQubit(Device):
 
         program.add_transform(convert_to_mbqc_gateset)
         program.add_transform(combine_global_phases)
-
         program.add_transform(split_at_non_clifford_gates)
         program.add_transform(convert_to_mbqc_formalism)
 
@@ -103,6 +99,7 @@ class FTQCQubit(Device):
         # # set up for backend execution (including MCM handling)
         if self.backend.diagonalize_mcms:
             program.add_transform(diagonalize_mcms)
+
         program.add_transform(
             qml.devices.preprocess.mid_circuit_measurements,
             device=self,
@@ -114,7 +111,6 @@ class FTQCQubit(Device):
         program.add_transform(
             qml.devices.preprocess.decompose,
             stopping_condition=stopping_condition_shots,
-            # stopping_condition_shots=stopping_condition_shots,
             skip_initial_state_prep=True,
             name=self.name,
         )
@@ -144,13 +140,17 @@ class FTQCQubit(Device):
         if config is None:
             config = ExecutionConfig()
 
-        # get mcm method - "device" if its an option, otherwise "one-shot"
-        default_mcm_method = _default_mcm_method(self.backend.capabilities, shots_present=True)
-        new_mcm_config = replace(config.mcm_config, mcm_method=default_mcm_method)
-        config = replace(config, mcm_config=new_mcm_config)
+        if config.mcm_config.mcm_method is None:
+            # get mcm method - "device" if its an option, otherwise "one-shot"
+            default_mcm_method = _default_mcm_method(self.backend.capabilities, shots_present=True)
+            new_mcm_config = replace(config.mcm_config, mcm_method=default_mcm_method)
+            config = replace(config, mcm_config=new_mcm_config)
         validate_mcm_method(
             self.backend.capabilities, config.mcm_config.mcm_method, shots_present=True
         )
+
+        if config.mcm_config.mcm_method == "deferred":
+            raise ValueError("Please no deferred measurements, thanks")
 
         return config
 
@@ -242,6 +242,7 @@ class LightningQubitBackend:
         # pylint: disable=protected-access
         for _ in range(sequence.shots.total_shots):
             self.device._statevector.reset_state()
+
             for segment in sequence.intermediate_tapes:
 
                 _ = self.simulate(
@@ -338,190 +339,6 @@ class NullQubitBackend:
             else:
                 raise RuntimeError("That's not a QuantumScript or a QuantumScriptSequence")
         return tuple(results)
-
-
-class QuantumScriptSequence:
-    """A sequence of tapes meant to be executed in order without resetting the system state.
-    Intermediate tapes may return mid-circuit measurements, or nothing. This is not currently
-    validated. The final tape returns terminal measurements."""
-
-    shortname = "QuantumScriptSequence"
-
-    def __init__(self, tapes, shots=None):
-
-        if shots is None:
-            shots = [tape.shots for tape in tapes]
-            if len(set(shots)) != 1:
-                raise RuntimeError(
-                    "All scripts in a QuantumScriptSequence must have the same shots"
-                )
-            shots = shots[0]
-        else:
-            if not isinstance(shots, Shots):
-                shots = Shots(shots)
-        self._shots = shots
-
-        self._tapes = []
-
-        for tape in tapes:
-            aux_tape = qml.tape.QuantumScript(
-                tape.operations,
-                tape.measurements,
-                shots=[1],
-            )
-            self._tapes.append(aux_tape)
-
-    @property
-    def tapes(self):
-        """Returns all the tapes in the sequence.
-
-        Returns:
-            list[QuantumScript]: list of all tapes
-        """
-        return self._tapes
-
-    @property
-    def final_tape(self):
-        """Returns the final tape in the sequence.
-
-        Returns:
-            QuantumScript: the final tape in the sequence, with terminal measurements
-        """
-        return self._tapes[-1]
-
-    @property
-    def intermediate_tapes(self):
-        """Returns all but the final tape in the sequence.
-
-        Returns:
-            list[QuantumScript]: list of all tapes except the tape with terminal measurements
-        """
-
-        return self._tapes[:-1]
-
-    @property
-    def measurements(self):
-        """Returns the final measurements for the sequence.
-
-        Returns:
-            list[.MeasurementProcess]: list of measurement processes
-        """
-        return self.final_tape.measurements
-
-    @property
-    def intermediate_measurements(self):
-        """Returns the intermediate measurements for all but the final tape. Since these
-        are in the middle of an execution, they are expected to be empty, or to be mid-circuit
-        measurements.
-
-        Returns:
-            list[list[MidMeasureMP]]: nested list of the returned MCMs for all but the final tape
-        """
-        return [tape.measurements for tape in self.intermediate_tapes]
-
-    @property
-    def operations(self):
-        """Returns the operations for each tape
-
-        Returns:
-            list[list[Operation]]: a nested list of the operations for each tape
-        """
-        return [tape.operations for tape in self.tapes]
-
-    @cached_property
-    def wires(self) -> qml.wires.Wires:
-        """Returns the wires used in the quantum script process
-
-        Returns:
-            ~.Wires: wires in quantum script process
-        """
-        wires = self.tapes[0].wires
-        for tape in self.tapes[1:]:
-            wires += tape.wires
-        return wires
-
-    @property
-    def num_wires(self) -> int:
-        """Returns the number of wires in the quantum script process
-
-        Returns:
-            int: number of wires in quantum script process
-        """
-        return len(self.wires)
-
-    @property
-    def shots(self):
-        """Returns a ``Shots`` object containing information about the number
-        and batches of shots
-
-        Returns:
-            ~.Shots: Object with shot information
-        """
-        return self._shots
-
-    def __repr__(self) -> str:
-        return f"<{self.shortname}: wires={list(self.wires)}>"
-
-    def map_to_standard_wires(self) -> "QuantumScriptSequence":
-        """Wrapper to apply qscript.map_to_standard_wires to each segment contained in the Sequence"""
-        wire_map = self._get_standard_wire_map()
-        if wire_map is None:
-            return self
-        new_tapes = []
-        for tape in self.tapes:
-            tapes, fn = qml.map_wires(tape, wire_map)
-            new_tapes.append(fn(tapes))
-
-        return self.copy(tapes=new_tapes)
-
-    def _get_standard_wire_map(self) -> dict:
-        """Helper function to produce the wire map for map_to_standard_wires. Wire map
-        is the same as if the sequence were a flat tape"""
-        flat_ops = []
-        for ops in self.operations:
-            flat_ops.extend(ops)
-
-        as_tape = qml.tape.QuantumScript(flat_ops, self.measurements)
-        return as_tape._get_standard_wire_map()  # pylint: disable=protected-access
-
-    def copy(self, copy_operations: bool = False, **update):
-        """Make it copy-able as if it were a tape where possible. Do not allow
-        modifications to operations or trainable parameters, because any transform
-        or function modifying operations on a tape will not work on a sequence of
-        tapes. Allow updating tapes as a whole as an alternative for
-        QuantumScriptSquence-specific functions to deal with modifying operations
-        on tapes.
-
-        This is not able to support trainable parameters and almost certainly also
-        has other flaws. It is not a thorough implementation of the desired behaviour."""
-        if copy_operations is True:
-            raise RuntimeError("Can't use copy_operations when copying a QuantumScriptSequence")
-
-        if update:
-            if "ops" in update:
-                update["operations"] = update["ops"]
-            for k in update:
-                if k not in ["tapes", "measurements", "shots"]:
-                    raise TypeError(f"{self.__class__}.copy() cannot update '{k}'")
-            if "tapes" in update and "measurements" in update:
-                raise RuntimeError(
-                    "Can't update tapes and measurements at the same time, as tapes include measurements"
-                )
-
-        _tapes = update.get("tapes", [copy.copy(tape) for tape in self.tapes])
-        _shots = update.get("shots", self.shots)
-
-        if "measurements" in update:
-            old_final_tape = _tapes.pop()
-            _new_final_tape = old_final_tape.copy(measurements=update["measurements"])
-            _tapes.append(_new_final_tape)
-
-        new_sequence = QuantumScriptSequence(
-            tapes=_tapes,
-            shots=_shots,
-        )
-
-        return new_sequence
 
 
 from pennylane.devices.preprocess import null_postprocessing
